completness tab

+++ dhis-2/dhis-web/dhis-web-apps/src/main/webapp/dhis-web-ndp-target/components/dataentry/dataentry-controller.js
@@ -30,6 +30,9 @@ ndpDataEntry.controller('DataEntryController',
         baseLineTargetActualDimensions: [],
         selectedAttributeCategoryCombo: null,
         selectedAttributeOptionCombos: null,
+        baseline: null,
+        target: null,
+        actual: null,
         dataSetsById: {},
         categoryCombosById: {},
         optionSets: [],
@@ -75,6 +78,7 @@ ndpDataEntry.controller('DataEntryController',
         $scope.dataValues = {};
         $scope.dataValuesCopy = {};
         $scope.saveStatus = {};
+        $scope.model.dataSetCompleteness = false;
     };
 
     //load datasets associated with the selected org unit.
@@ -96,6 +100,7 @@ ndpDataEntry.controller('DataEntryController',
         $scope.model.periodOffset = 0;
         $scope.dataValues = {};
         $scope.dataValuesCopy = {};
+        $scope.model.dataSetCompleteness = false;
         if( angular.isObject($scope.model.selectedDataSet) && $scope.model.selectedDataSet.id){
             $scope.loadDataSetDetails();
         }
@@ -105,6 +110,7 @@ ndpDataEntry.controller('DataEntryController',
         $scope.dataValues = {};
         $scope.dataValuesCopy = {};
         $scope.saveStatus = {};
+        $scope.model.dataSetCompleteness = false;
         $scope.loadDataEntryForm();
     });
 
@@ -127,6 +133,18 @@ ndpDataEntry.controller('DataEntryController',
                 CommonUtils.notify('error', 'missing_dataset_category_combo');
                 return;
             }
+
+            angular.forEach($scope.model.selectedAttributeCategoryCombo.categoryOptionCombos, function(aoc){
+                if ( aoc.btaDimensionType === 'baseline' ){
+                    $scope.model.baseline = aoc;
+                }
+                else if ( aoc.btaDimensionType === 'target' ){
+                    $scope.model.target = aoc;
+                }
+                else if ( aoc.btaDimensionType === 'actual' ){
+                    $scope.model.actual = aoc;
+                }
+            });
 
             $scope.model.selectedDataSet.dataElements = orderByFilter( $scope.model.selectedDataSet.dataElements, '-displayName').reverse();
             $scope.model.dataElements = [];
@@ -201,7 +219,15 @@ ndpDataEntry.controller('DataEntryController',
                         response.completeDataSetRegistrations.length &&
                         response.completeDataSetRegistrations.length > 0){
 
-                    $scope.model.dataSetCompletness = true;
+                    var cdsrs = [];
+                    angular.forEach(response.completeDataSetRegistrations, function(cdsr){
+                        if ( cdsr.attributeOptionCombo === $scope.model.baseline.id || cdsr.attributeOptionCombo === $scope.model.target.id ){
+                            cdsrs.push( cdsr);
+                        }
+                    });
+                    if ( cdsrs.length > 0 ){
+                        $scope.model.dataSetCompleteness = true;
+                    }
                 }
             });
         }
@@ -310,27 +336,41 @@ ndpDataEntry.controller('DataEntryController',
         });
     };
 
+    $scope.isDisabled = function( aoc ){
+        return !aoc.dWrite || $scope.model.dataSetCompleteness;
+    };
+
+    $scope.canSubmitData = function(){
+        return CommonUtils.userHasAuthority('F_COMPLETE_DATASET');
+    };
+
+    $scope.canUnSubmitData = function(){
+        return CommonUtils.userHasAuthority('F_UNCOMPLETE_DATASET');
+    };
+
     $scope.saveCompletness = function(){
         var modalOptions = {
             closeButtonText: 'no',
             actionButtonText: 'yes',
-            headerText: 'mark_complete',
-            bodyText: 'are_you_sure_to_save_completeness'
+            headerText: 'submit_data',
+            bodyText: 'are_you_sure_to_submit_data'
         };
 
         ModalService.showModal({}, modalOptions).then(function(result){
             var dsr = {completeDataSetRegistrations: []};
             angular.forEach($scope.model.selectedAttributeCategoryCombo.categoryOptionCombos, function(aoc){
-                dsr.completeDataSetRegistrations.push( {dataSet: $scope.model.selectedDataSet.id, organisationUnit: $scope.selectedOrgUnit.id, period: $scope.model.selectedPeriod.id, attributeOptionCombo: aoc.id} );
+                if ( aoc.btaDimensionType === 'baseline'  || aoc.btaDimensionType === 'target' ){
+                    dsr.completeDataSetRegistrations.push( {dataSet: $scope.model.selectedDataSet.id, organisationUnit: $scope.selectedOrgUnit.id, period: $scope.model.selectedPeriod.id, attributeOptionCombo: aoc.id} );
+                }
             });
 
             CompletenessService.saveDsr(dsr).then(function(response){
                 var dialogOptions = {
                     headerText: 'success',
-                    bodyText: 'marked_complete'
+                    bodyText: 'data_submitted'
                 };
                 DialogService.showDialog({}, dialogOptions);
-                $scope.model.dataSetCompletness = true;
+                $scope.model.dataSetCompleteness = true;
 
             }, function(response){
                 CommonUtils.errorNotifier( response );
@@ -342,30 +382,32 @@ ndpDataEntry.controller('DataEntryController',
         var modalOptions = {
             closeButtonText: 'no',
             actionButtonText: 'yes',
-            headerText: 'mark_incomplete',
-            bodyText: 'are_you_sure_to_delete_completeness'
+            headerText: 'un_submit_data',
+            bodyText: 'are_you_sure_to_recall_data_and_edit'
         };
 
         ModalService.showModal({}, modalOptions).then(function(result){
 
             angular.forEach($scope.model.selectedAttributeCategoryCombo.categoryOptionCombos, function(aoc){
-                CompletenessService.delete($scope.model.selectedDataSet.id,
-                    $scope.model.selectedPeriod.id,
-                    $scope.selectedOrgUnit.id,
-                    $scope.model.selectedAttributeCategoryCombo.id,
-                    CommonUtils.getOptionIds(aoc.categoryOptions),
-                    false).then(function(response){
+                if ( aoc.btaDimensionType === 'baseline'  || aoc.btaDimensionType === 'target' ){
+                    CompletenessService.delete($scope.model.selectedDataSet.id,
+                        $scope.model.selectedPeriod.id,
+                        $scope.selectedOrgUnit.id,
+                        $scope.model.selectedAttributeCategoryCombo.id,
+                        CommonUtils.getOptionIds(aoc.categoryOptions),
+                        false).then(function(response){
 
-                    var dialogOptions = {
-                        headerText: 'success',
-                        bodyText: 'marked_incomplete'
-                    };
-                    DialogService.showDialog({}, dialogOptions);
-                    $scope.model.dataSetCompletness = false;
+                        var dialogOptions = {
+                            headerText: 'success',
+                            bodyText: 'data_unsubmitted'
+                        };
+                        DialogService.showDialog({}, dialogOptions);
+                        $scope.model.dataSetCompleteness = false;
 
-                }, function(response){
-                    CommonUtils.errorNotifier( response );
-                });
+                    }, function(response){
+                        CommonUtils.errorNotifier( response );
+                    });
+                }
             });
         });
     };
